- Marten fixed trac #218, ambiguous datatypes in lighter2.
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@26081 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ apps/tools/lighter2/directlight.cpp
@@ -305,7 +305,7 @@ namespace lighter
           ElementProxy ep = prim.GetElement (eidx);
           csColor c = lmElementShader (sector, ep, masterSampler, rt);
 
-          uint u, v;
+          size_t u, v;
           prim.GetElementUV (eidx, u, v);
           u += minUV.x;
           v += minUV.y;
